refactor: spilit utils file
