+++ src/utils/dom.ts
@@ -1,0 +1,3 @@
+export function isVerticalScrollable(): boolean {
+  return document.body.scrollHeight > document.body.clientHeight
+}
