+++ src/utils/number.ts
@@ -1,0 +1,28 @@
+export function getRandomNumber(min: number, max: number): number {
+  return parseInt((Math.random() * (max - min) + min).toString(), 10)
+}
+
+export function getRandomBool(): boolean {
+  return Math.random() > 0.5
+}
+
+export function roundDown(n: number, decimals = 1): number {
+  return Math.floor(n * Math.pow(10, decimals)) / Math.pow(10, decimals)
+}
+
+const MILLION = 1e6
+
+export const formatCount = (count: number, lang: string): number | string => {
+  if (lang === 'zh') {
+    if (count >= MILLION) {
+      return `${roundDown(count / MILLION)} 百万`
+    } else if (count >= 10000) {
+      return `${roundDown(count / 10000)} 万`
+    }
+    return count
+  }
+  if (count >= MILLION) {
+    return `${roundDown(count / MILLION)}m`
+  }
+  return count >= 1000 ? `${roundDown(count / 1000)}k` : count
+}
